+++ src/ft_stdlib/ft_ultoa.c
@@ -11,8 +11,8 @@ char *ft_ultoa(unsigned long number, char *buffer, int base) {
   if (!number)
     *buffer++ = BASE[0];
   while (number) {
-    *buffer++ = BASE[number % base];
-    number /= base;
+    *buffer++ = BASE[number % (unsigned long)base];
+    number /= (unsigned long)base;
   }
   *buffer = 0;
   return ft_strrev(head);
