+++ src/ft_string/ft_strlen.c
@@ -1,9 +1,9 @@
-#include <stddef.h>
+#include "ft_string.h"
 
 size_t ft_strlen(char const *str) {
   char const *head = str;
 
   while (*str)
     ++str;
-  return str - head;
+  return (size_t)(str - head);
 }
